requirements.txt
langchain-openai
langchain-core
python-dotenv
streamlit
langchain-community
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
faiss-cpu
wikipedia
beautifulsoup4
cassio
arxiv
langchainhub
langchain
groq
langchain-groq
sentence_transformers
